refactor parseQuery in both tree & graph
